+++ stubs/defaultConfig.stub.js
@@ -392,6 +392,11 @@ module.exports = {
       '11': '11',
       '12': '12',
     },
+    outline: {
+      none: ['2px solid transparent', '2px'],
+      white: ['2px dotted white', '2px'],
+      black: ['2px dotted black', '2px'],
+    },
     padding: theme => theme('spacing'),
     placeholderColor: theme => theme('colors'),
     placeholderOpacity: theme => theme('opacity'),